+++ src/features/Conversation/Messages/Tool/index.tsx
@@ -1,17 +1,28 @@
-import { memo, useState } from 'react';
-import { Flexbox } from 'react-layout-kit';
+import { Icon } from '@lobehub/ui';
+import { ConfigProvider, Empty } from 'antd';
+import { useTheme } from 'antd-style';
+import { LucideSquareArrowLeft, LucideSquareArrowRight } from 'lucide-react';
+import { memo, useContext, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Center, Flexbox } from 'react-layout-kit';
 
+import PluginRender from '@/features/PluginsUI/Render';
 import { useChatStore } from '@/store/chat';
-import { chatSelectors } from '@/store/chat/selectors';
+import { chatPortalSelectors, chatSelectors } from '@/store/chat/selectors';
 import { ChatMessage } from '@/types/message';
 
-import PluginRender from '../../Plugins/Render';
 import Arguments from '../components/Arguments';
 import Inspector from './Inspector';
 
 export const ToolMessage = memo<ChatMessage>(({ id, content, pluginState, plugin }) => {
-  const loading = useChatStore(chatSelectors.isPluginApiInvoking(id));
+  const [loading, isMessageToolUIOpen] = useChatStore((s) => [
+    chatSelectors.isPluginApiInvoking(id)(s),
+    chatPortalSelectors.isArtifactMessageUIOpen(id)(s),
+  ]);
+  const { direction } = useContext(ConfigProvider.ConfigContext);
+  const { t } = useTranslation('plugin');
 
+  const theme = useTheme();
   const [showRender, setShow] = useState(plugin?.type !== 'default');
 
   return (
@@ -19,14 +30,30 @@ export const ToolMessage = memo<ChatMessage>(({ id, content, pluginState, plugin
       <Inspector
         arguments={plugin?.arguments}
         content={content}
+        id={id}
         identifier={plugin?.identifier}
         loading={loading}
         payload={plugin}
         setShow={setShow}
         showRender={showRender}
       />
-      {showRender || loading ? (
+      {isMessageToolUIOpen ? (
+        <Center paddingBlock={8} style={{ background: theme.colorFillQuaternary, borderRadius: 4 }}>
+          <Empty
+            description={t('showInPortal')}
+            image={
+              <Icon
+                color={theme.colorTextQuaternary}
+                icon={direction === 'rtl' ? LucideSquareArrowLeft : LucideSquareArrowRight}
+                size={'large'}
+              />
+            }
+            imageStyle={{ height: 24 }}
+          />
+        </Center>
+      ) : showRender || loading ? (
         <PluginRender
+          arguments={plugin?.arguments}
           content={content}
           id={id}
           identifier={plugin?.identifier}